+++ src/config/index.ts
@@ -45,7 +45,7 @@ async function getBranchConfigFiles() {
     return branchConfigFiles;
 }
 
-export const getConfig = async (layer: string = 'user'): Promise<any> => {
+export const getConfig = async (layer = 'user'): Promise<any> => {
     const defaultConfig = await loadFromConfigFile(projectConfigFiles);
     if (layer === 'project') {
         return defaultConfig;